Fully abstracted OpenGL rendering from application.cpp by improving Renderer class and introducing RenderCommand, RendererAPI and OpenGL specific renderer API.

+++ Rise/src/Platform/OpenGL/OpenGLBuffer.cpp
@@ -35,7 +35,7 @@ namespace Rise
 		glCreateBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
 		// Pass our drawing order into buffer. Static draw since it wont change at runtime.
-		glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t), indices, GL_STATIC_DRAW);
+		glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<long long>(count * sizeof(uint32_t)), indices, GL_STATIC_DRAW);
 	}
 
 	OpenGLIndexBuffer::~OpenGLIndexBuffer()
